refactor: straight cleanup / modularizing
