+++ public/app/utils.ts
@@ -1,0 +1,4 @@
+export const tapl = <T>(
+  x: T,
+  fmt: (x: T) => string,
+) => (console.log(fmt(x)), x);
